% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{cluster_clonotypes}
\alias{cluster_clonotypes}
\title{k-means clustering on clonotype features}
\usage{
cluster_clonotypes(clono_features, k = 6)
}
\arguments{
\item{clono_features}{summarise_clonotypes() 的输出。}

\item{k}{聚类数。}
}
\value{
增加 cluster 列的 data.frame。
}
\description{
k-means clustering on clonotype features
}
