% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repertoire_summary.R
\name{getCloneAbundanceDistribution}
\alias{getCloneAbundanceDistribution}
\title{获取克隆丰度向量}
\usage{
getCloneAbundanceDistribution(df)
}
\arguments{
\item{df}{单个样本 AIRR 数据。}
}
\value{
各克隆的 size 向量。
}
\description{
获取克隆丰度向量
}
