% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repertoire_summary.R
\name{getLengthDistribution}
\alias{getLengthDistribution}
\title{获取 CDR3 长度分布}
\usage{
getLengthDistribution(df, type = "aa")
}
\arguments{
\item{df}{单个样本 AIRR 数据。}

\item{type}{"aa" 表示氨基酸长度，"nt" 表示核苷酸长度。}
}
\value{
整数向量，长度分布。
}
\description{
获取 CDR3 长度分布
}
