% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repertoire_summary.R
\name{getVGeneDistributions}
\alias{getVGeneDistributions}
\title{计算 V 基因使用频率}
\usage{
getVGeneDistributions(df)
}
\arguments{
\item{df}{单个样本 AIRR 数据。}
}
\value{
data.table，包含 \code{gene} 和 \code{frequency} 列。
}
\description{
计算 V 基因使用频率
}
