% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repertoire_summary.R
\name{plotCloneAbundance}
\alias{plotCloneAbundance}
\title{绘制克隆丰度 rank–abundance 曲线}
\usage{
plotCloneAbundance(df, title = "Clonal Expansion Rank–Abundance Curve")
}
\arguments{
\item{df}{单个样本 AIRR 数据。}

\item{title}{图标题。}
}
\description{
绘制克隆丰度 rank–abundance 曲线
}
