% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repertoire_summary.R
\name{plotLengthDistribution}
\alias{plotLengthDistribution}
\title{绘制 CDR3 长度分布直方图}
\usage{
plotLengthDistribution(df, type = "aa", title = "CDR3 Length Distribution")
}
\arguments{
\item{df}{单个样本 AIRR 数据。}

\item{type}{"aa" 或 "nt"。}

\item{title}{图标题。}
}
\description{
绘制 CDR3 长度分布直方图
}
