% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repertoire_summary.R
\name{plotVJPairing}
\alias{plotVJPairing}
\title{绘制 V–J 配对热图}
\usage{
plotVJPairing(df, top_v = 20, top_j = 35, title = "V–J Gene Pairing Landscape")
}
\arguments{
\item{df}{单个样本 AIRR 数据。}

\item{top_v}{保留使用频率最高的 V 个数。}

\item{top_j}{保留使用频率最高的 J 个数。}

\item{title}{图标题。}
}
\description{
绘制 V–J 配对热图
}
