% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShieldAIRR.R
\docType{package}
\name{ShieldAIRR-package}
\alias{ShieldAIRR}
\alias{ShieldAIRR-package}
\title{ShieldAIRR: TCR/BCR AIRR Repertoire Analysis Toolkit}
\description{
ShieldAIRR 提供了一系列面向 AIRR 格式 AIRR 数据的高层封装函数：
}
\details{
\itemize{
\item 单样本整体概览：\code{\link{summarizeRepertoirePlot}}
\item CDR3 理化性质景观：\code{\link{shield_cdr3_landscape}}
\item 时间序列克隆轨迹聚类：\code{\link{make_long}},
\code{\link{summarise_clonotypes}},
\code{\link{cluster_clonotypes}},
\code{\link{plot_cluster_traj}}
\item V/J 基因差异使用分析：\code{\link{shield_vj_summarise}},
\code{\link{shield_vj_deseq}}
}
}
\author{
\strong{Maintainer}: XF Zhang \email{cheungxuefei@gmail.com}

}
\keyword{internal}
