% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{summarise_clonotypes}
\alias{summarise_clonotypes}
\title{Time-series: summarise clonotype features}
\usage{
summarise_clonotypes(long)
}
\arguments{
\item{long}{make_long() 的输出}
}
\value{
data.frame，每个 clonotype 对应一行
}
\description{
Time-series: summarise clonotype features
}
