% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{make_long}
\alias{make_long}
\title{Time-series: make long-format clonotype table}
\usage{
make_long(
  dfs,
  clonotype_col = "junction_aa",
  abundance_col = "duplicate_count",
  min_count = 1
)
}
\arguments{
\item{dfs}{命名列表，每个元素是一个 data.frame（一个时间点的 AIRR 数据）}

\item{clonotype_col}{定义克隆的列名，默认 "junction_aa"}

\item{abundance_col}{丰度列名，默认 "duplicate_count"}

\item{min_count}{保留的最小丰度阈值}
}
\value{
data.frame，包含 junction_aa, duplicate_count, time, l_abundance
}
\description{
Time-series: make long-format clonotype table
}
