% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series.R
\name{plot_cluster_traj}
\alias{plot_cluster_traj}
\title{Time-series: plot cluster trajectories}
\usage{
plot_cluster_traj(long, clono_clustered, k)
}
\arguments{
\item{long}{make_long() 的输出}

\item{clono_clustered}{cluster_clonotypes() 的输出}

\item{k}{聚类簇数}
}
\description{
Time-series: plot cluster trajectories
}
