% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repertoire_summary.R
\name{summarizeRepertoirePlot}
\alias{summarizeRepertoirePlot}
\title{单样本 TCR/BCR repertoire 综合概览图}
\usage{
summarizeRepertoirePlot(df, sample_name = "Sample", output_pdf = TRUE)
}
\arguments{
\item{df}{单个样本 AIRR 数据。}

\item{sample_name}{样本名称，用于标题。}

\item{output_pdf}{是否自动保存 PDF/PNG 文件。}
}
\value{
不可见返回 patchwork 对象 \code{final_plot}。
}
\description{
单样本 TCR/BCR repertoire 综合概览图
}
